requirements_docker.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
requests>=2.28.0
gradio>=3.40.0
omegaconf>=2.3.0
einops>=0.6.1
opencv-python>=4.7.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
safetensors>=0.3.1
peft>=0.4.0
